@@ -40,8 +40,9 @@ SELF_CLOSING = {
 
 def make(
     tag: str,
-    attrs: t.Mapping[str, t.Optional[str]],
-    close: t.Optional[bool] = None,
+    attrs: t.Mapping[str, str | None],
+    *,
+    close: bool | None = None,
 ) -> str:
     """
     Helper for quickly constructing a HTML tag.
@@ -62,20 +63,20 @@ def make(
 
 @dataclasses.dataclass
 class Element:
-    tag: t.Optional[str]
+    tag: str | None
     attrs: dict = dataclasses.field(default_factory=dict)
     children: list = dataclasses.field(default_factory=list)
 
     def __getitem__(self, i: int):
         return self.children[i]
 
-    def __len__(self):
+    def __len__(self) -> int:
         return len(self.children)
 
     def __iter__(self):
         return iter(self.children)
 
-    def serialize(self, dest: t.Optional[io.TextIOBase] = None) -> io.TextIOBase:
+    def serialize(self, dest: io.TextIOBase | None = None) -> io.TextIOBase:
         if dest is None:
             dest = io.StringIO()
         if self.tag is not None:
@@ -101,7 +102,7 @@ class Parser(HTMLParser):
     Parses a HTML document into
     """
 
-    def __init__(self):
+    def __init__(self) -> None:
         super().__init__()
         self.root = Element(tag=None)
         self.stack = [self.root]
@@ -110,28 +111,28 @@ class Parser(HTMLParser):
     def top(self):
         return self.stack[-1]
 
-    def handle_starttag(self, tag, attrs):
+    def handle_starttag(self, tag, attrs) -> None:
         elem = Element(tag=tag, attrs=dict(attrs))
         self.top.children.append(elem)
         if tag not in SELF_CLOSING:
             self.stack.append(elem)
 
-    def handle_startendtag(self, tag, attrs):
+    def handle_startendtag(self, tag, attrs) -> None:
         elem = Element(tag=tag, attrs=dict(attrs))
         self.top.children.append(elem)
 
-    def handle_endtag(self, tag):
+    def handle_endtag(self, tag) -> None:
         if tag not in SELF_CLOSING:
             while len(self.stack) > 1:
                 self.stack.pop()
                 if tag == self.top.tag:
                     break
 
-    def handle_data(self, data):
+    def handle_data(self, data) -> None:
         if data != "":
             self.top.children.append(data)
 
-    def error(self, message: str):
+    def error(self, message: str) -> None:
         # This method is undocumented in HTMLParser, but pylint is moaning
         # about it, so...
         logger.error("Error in Parser: %s", message)  # pragma: no cover
