@@ -9,18 +9,18 @@ class Actions:
     Batches together a set of actions to take.
     """
 
-    def __init__(self):
+    def __init__(self) -> None:
         super().__init__()
         self.actions = []
 
-    def _append(self, **params):
+    def _append(self, **params) -> None:
         for key, value in list(params.items()):
             if not value:
                 del params[key]
             elif isinstance(value, datetime.datetime):
                 # Treat naive timezones as UTC for sanity's sake.
                 if value.tzinfo is None:
-                    value = value.replace(tzinfo=datetime.timezone.utc)
+                    value = value.replace(tzinfo=datetime.UTC)
                 params[key] = int(value.timestamp())
             elif key == "tags":
                 params[key] = ",".join(value)
@@ -29,37 +29,37 @@ class Actions:
     def add(
         self,
         url: str,
-        title: t.Optional[str] = None,
+        title: str | None = None,
         tags: t.Sequence[str] = (),
-        ref_id: t.Optional[str] = None,
-    ):
+        ref_id: str | None = None,
+    ) -> None:
         self._append(action="add", url=url, title=title, tags=tags, ref_id=ref_id)
 
-    def archive(self, item_id: str):
+    def archive(self, item_id: str) -> None:
         self._append(action="archive", item_id=item_id)
 
-    def delete(self, item_id: str):
+    def delete(self, item_id: str) -> None:
         self._append(action="delete", item_id=item_id)
 
-    def favorite(self, item_id: str):
+    def favorite(self, item_id: str) -> None:
         self._append(action="favorite", item_id=item_id)
 
-    def unfavorite(self, item_id: str):
+    def unfavorite(self, item_id: str) -> None:
         self._append(action="unfavorite", item_id=item_id)
 
-    def tags_add(self, item_id: str, tags: t.Sequence[str]):
+    def tags_add(self, item_id: str, tags: t.Sequence[str]) -> None:
         self._append(action="tags_add", item_id=item_id, tags=tags)
 
-    def tags_remove(self, item_id: str, tags: t.Sequence[str]):
+    def tags_remove(self, item_id: str, tags: t.Sequence[str]) -> None:
         self._append(action="tags_remove", item_id=item_id, tags=tags)
 
-    def tags_replace(self, item_id: str, tags: t.Sequence[str]):
+    def tags_replace(self, item_id: str, tags: t.Sequence[str]) -> None:
         self._append(action="tags_replace", item_id=item_id, tags=tags)
 
-    def tags_clear(self, item_id: str):
+    def tags_clear(self, item_id: str) -> None:
         self._append(action="tags_clear", item_id=item_id)
 
-    def tag_rename(self, old_tag: str, new_tag: str):
+    def tag_rename(self, old_tag: str, new_tag: str) -> None:
         self._append(action="tag_rename", old_tag=old_tag, new_tag=new_tag)
 
 
@@ -70,12 +70,12 @@ class Client:
 
     PREFIX = "https://getpocket.com/v3/"
 
-    def __init__(self, consumer_key: str, access_token: t.Optional[str] = None):
+    def __init__(self, consumer_key: str, access_token: str | None = None) -> None:
         super().__init__()
         self._consumer_key = consumer_key
         self.access_token = access_token
 
-    def make_request(self, method: str, params: t.Optional[dict] = None):
+    def make_request(self, method: str, params: dict | None = None):
         data = {"consumer_key": self._consumer_key}
         if self.access_token is not None:
             data["access_token"] = self.access_token
@@ -95,7 +95,7 @@ class Client:
         with request.urlopen(req) as response:
             return json.load(response)
 
-    def oauth_request(self, redirect_uri: str, state: t.Optional[str] = None):
+    def oauth_request(self, redirect_uri: str, state: str | None = None):
         params = {"redirect_uri": redirect_uri}
         if state is not None:
             params["state"] = state
@@ -107,9 +107,9 @@ class Client:
     def add(
         self,
         url: str,
-        title: t.Optional[str] = None,
+        title: str | None = None,
         tags: t.Sequence[str] = (),
-        tweet_id: t.Optional[str] = None,
+        tweet_id: str | None = None,
     ):
         params = {
             "url": url,
@@ -117,7 +117,7 @@ class Client:
         }
         for key, value in [("title", title), ("tweet_id", tweet_id)]:
             if value is not None:
-                params[key] = value
+                params[key] = value  # noqa: PERF403
         return self.make_request("add", params)
 
     def modify(self, actions: Actions):
