@@ -1,12 +1,11 @@
-import unittest
 
 from komorebi import formatting
 
 
-class TestMarkdown(unittest.TestCase):
-    def test_empty(self):
-        self.assertEqual(formatting.render_markdown(None), "")
-        self.assertEqual(formatting.render_markdown(""), "")
+def test_empty():
+    assert formatting.render_markdown(None) == ""
+    assert formatting.render_markdown("") == ""
 
-    def test_basic_rendering(self):
-        self.assertEqual(formatting.render_markdown("foo"), "<p>foo</p>")
+
+def test_basic_rendering():
+    assert formatting.render_markdown("foo") == "<p>foo</p>"
